[IMP] base_changeset: Refactor code (reduce code)
TT45803
