+++ base_changeset/static/src/js/form_backend.js
@@ -3,8 +3,6 @@ odoo.define("base_changeset.form_backend", function (require) {
 
     var FormRenderer = require("web.FormRenderer");
     var FormController = require("web.FormController");
-    var core = require("web.core");
-    var qweb = core.qweb;
 
     FormController.include({
         start: function () {
@@ -12,7 +10,6 @@ odoo.define("base_changeset.form_backend", function (require) {
                 .apply(this, arguments)
                 .then(this._updateChangeset.bind(this));
         },
-
         update: function () {
             var self = this;
             var res = this._super.apply(this, arguments);
@@ -21,7 +18,6 @@ odoo.define("base_changeset.form_backend", function (require) {
             });
             return res;
         },
-
         _updateChangeset: function () {
             var self = this;
             var state = this.model.get(this.handle);
@@ -31,16 +27,13 @@ odoo.define("base_changeset.form_backend", function (require) {
                     self.renderer.renderChangesetPopovers(changeset);
                 });
         },
-
         applyChange: function (id) {
             this.model.applyChange(id).then(this.reload.bind(this));
         },
-
         rejectChange: function (id) {
             this.model.rejectChange(id).then(this.reload.bind(this));
         },
     });
-
     FormRenderer.include({
         renderChangesetPopovers: function (changeset) {
             var self = this;
@@ -63,55 +56,6 @@ odoo.define("base_changeset.form_backend", function (require) {
                 }
                 self._renderChangesetPopover($label, changes);
             });
-        },
-
-        _renderChangesetPopover: function ($el, changes) {
-            var self = this;
-            if (this.mode !== "readonly") {
-                return;
-            }
-            var $button = $(
-                qweb.render("ChangesetButton", {
-                    count: changes.length,
-                })
-            );
-
-            $el.append($button);
-
-            var options = {
-                content: function () {
-                    var $content = $(
-                        qweb.render("ChangesetPopover", {
-                            changes: changes,
-                        })
-                    );
-                    $content.find(".base_changeset_apply").on("click", function () {
-                        self._applyClicked($(this));
-                    });
-                    $content.find(".base_changeset_reject").on("click", function () {
-                        self._rejectClicked($(this));
-                    });
-                    return $content;
-                },
-                html: true,
-                placement: "bottom",
-                title: "Pending Changes",
-                trigger: "focus",
-                delay: {show: 0, hide: 100},
-                template: qweb.render("ChangesetTemplate"),
-            };
-
-            $button.popover(options);
-        },
-
-        _applyClicked: function ($el) {
-            var id = parseInt($el.data("id"), 10);
-            this.getParent().applyChange(id);
-        },
-
-        _rejectClicked: function ($el) {
-            var id = parseInt($el.data("id"), 10);
-            this.getParent().rejectChange(id);
         },
     });
 });
